+++ templates/javascript/jsdoc/constants.jsdoc.js
@@ -1,0 +1,3 @@
+exports.SCOPE_TYPEDEF = 'typedef';
+
+exports.SCOPE_FUNC_PARAMS = 'func_params';