+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 CHANGELOG - generator-api-on-calling
 ---
 
+## 0.1.1
+
++ jsdoc requestBody
++ jsdoc responseBody
++ jsdoc types comments
+
+## 0.1.0
+
++ parameters jsdoc
+
+## 0.0.7
+
+- refactor(templates/javascript/template.js): rm useless keyword: SERVICE_NAMESPACE
+
+## 0.0.6
+
+- refactor(templates/javascript/template.js): rm useless data object
+
 ## 0.0.5
 
 - chore(package.json): files add common, CHANGELOG.md